Get basic scraper working
In this commit the scraper function is ready and reads a list of urls from the file list.js. This file should not be commited, since it might have secret information, so ther eis a list.js.dist to show how this file should be.
In order for the scrape function call another lambda function, I had to use the aws-sdk and call the function by its full name web-monitoring-dev-requestUrl. Right now this is hardcoded, but I will try to not have it hardcoded in case we need to switch environments (prod and dev, for example).