Split business logic from serverless
To make the functions more reusable if needed in the future, they will no longer depend on the event or context received by lambda. This way, the handler.js file will simply choose the needed information and call the specific function.
This is a big change because it is easier to write unit tests for those functions, since it is not coupled.